Split test_pagination.py file
